+++ packages/blockera/js/test/compatibility.blocksy.e2e.cy.js
@@ -70,4 +70,62 @@ describe('Compatibility with Blocksy Theme', () => {
 		// 		'color: var(--theme-link-initial-color)'
 		// 	);
 	});
+
+	it('Width Size variables', () => {
+		cy.getParentContainer('Width').within(() => {
+			cy.openValueAddon();
+		});
+
+		cy.get('.blockera-component-popover')
+			.last()
+			.within(() => {
+				cy.contains('Blocksy Width Size').should('exist');
+			});
+
+		// select variable
+		cy.selectValueAddonItem('normal-container-max-width');
+
+		cy.getIframeBody().within(() => {
+			cy.get('#blockera-styles-wrapper')
+				.invoke('text')
+				.should(
+					'include',
+					'width: var(--theme-normal-container-max-width)'
+				);
+		});
+
+		//Check store
+		getWPDataObject().then((data) => {
+			expect({
+				settings: {
+					name: 'Normal Container Max Width',
+					id: 'normal-container-max-width',
+					value: '1290px',
+					type: 'width-size',
+					group: 'blocksy-width-size',
+					var: '--theme-normal-container-max-width',
+					label: 'Normal Container Max Width',
+					reference: {
+						type: 'theme',
+						theme: 'blocksy',
+					},
+				},
+				name: 'Normal Container Max Width',
+				isValueAddon: true,
+				valueType: 'variable',
+			});
+
+			//Check frontend
+			// savePage();
+
+			// redirectToFrontPage();
+
+			// cy.get('style#blockera-inline-css')
+			// 	.invoke('text')
+			// 	.should(
+			// 		'include',
+			// 		'width: var(--theme-normal-container-max-width)'
+			// 	);
+		});
+	});
 });